+++ breeze-core/src/main/kotlin/domain/Inlineable.kt
@@ -1,6 +1,8 @@
 // Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
 // Breeze is blowing...
 
+@file:Suppress("METHOD_OF_ANY_IMPLEMENTED_IN_INTERFACE")
+
 package com.windea.breezeframework.core.domain
 
 /**
@@ -16,9 +18,9 @@ interface Inlineable : CharSequence {
 
 	override val length get() = inlineText.length
 
-	override fun get(index: Int) = inlineText[index]
+	 override fun get(index: Int) = inlineText[index]
 
 	override fun subSequence(startIndex: Int, endIndex: Int) = inlineText.subSequence(startIndex, endIndex)
 
-	override fun toString(): String
+	override fun toString(): String = ""
 }