+++ breeze-core/src/main/kotlin/domain/data/DataBuilder.kt
@@ -1,9 +1,6 @@
 // Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
 // Breeze is blowing...
 
-// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
-// Breeze is blowing...
-
 package com.windea.breezeframework.core.domain.data
 
 /**