+++ breeze-unstable/src/main/kotlin/core/domain/misc/DirectionX.kt
@@ -1,0 +1,12 @@
+// Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
+// Breeze is blowing...
+
+package com.windea.breezeframework.core.domain.misc
+
+/**
+ * 基于X轴的一维方向。
+ */
+enum class DirectionX {
+	Forward,
+	Backward
+}